log-systemd: fix condition
In the unlikely event that systemd contains a bug, sd_journal_get_data
succeeds, but returns truncated data. In that case, the length might
become negative. Detect this too.

+++ log-systemd.c
@@ -102,7 +102,7 @@ log_read_line(char *buf, size_t buf_size) {
 	d += sizeof("MESSAGE=") - 1;
 	l -= sizeof("MESSAGE=") - 1;
 
-	if (l <= 0) {
+	if ((int)l <= 0) {
 		/* empty message or systemd bug ? */
 		return 0;
 	}
